feat: 🎸 Added support for generating a tree with default data
It is now possible to create a new Tree with default data by passing an
array with JsonNode's. This is useful when you have a copy of your JSON
tree saved somewhere and you want to recreate it. A JSON representation
of the tree can be created by simple doing `tree.toJson()`.